requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
aiohttp-retry==2.4.6
antlr4-python3-runtime==4.8
appdirs==1.4.4
astunparse==1.6.3
async-timeout==3.0.1
atpublic==2.3
attrs==21.2.0
beautifulsoup4==4.9.3
black==21.4b2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
clang==5.0
click==8.0.1
cloudpickle==1.6.0
colorama==0.4.4
commonmark==0.9.1
configobj==5.0.6
cox==0.1.post3
cycler==0.10.0
Cython==0.29.23
decorator==4.4.2
dictdiffer==0.9.0
dill==0.3.4
diskcache==5.2.1
distro==1.6.0
dpath==2.0.5
dulwich==0.20.25
dvc==2.8.2
ffmpy==0.3.0
flatbuffers==1.12
flatten-dict==0.4.2
flufl.lock==6.0
fsspec==2021.10.1
ftfy==6.0.3
funcy==1.16
future==0.18.2
fvcore==0.1.5.post20210617
gast==0.4.0
gensim==4.0.1
gitdb==4.0.7
GitPython==3.1.24
google-api-core==1.29.0
google-api-python-client==2.7.0
google-auth==1.30.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
GPUtil==1.4.0
grandalf==0.6
grpcio==1.38.1
h5py==3.1.0
httplib2==0.19.1
hydra-core==1.1.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.4
importlib-resources==5.3.0
iopath==0.1.8
joblib==1.0.1
jsonpatch==1.32
jsonpointer==2.1
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
loky==2.9.0
mailchecker==4.0.13
Markdown==3.3.4
matplotlib==3.4.2
moviepy==1.0.3
multidict==5.2.0
mypy-extensions==0.4.3
nanotime==0.5.2
networkx==2.5.1
nltk==3.6.2
numexpr==2.7.3
oauth2client==4.1.3
oauthlib==3.1.1
omegaconf==2.1.0
opencv-python==4.5.2.52
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
pathspec==0.9.0
phonenumbers==8.12.36
Pillow==8.2.0
ply==3.11
portalocker==2.3.0
prettytable==2.1.0
proglog==0.1.9
protobuf==3.17.2
psutil==5.8.0
py3nvml==0.2.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
pydot==1.4.2
PyDrive==1.3.1
pygit2==1.7.0
Pygments==2.10.0
PyYAML==5.4.1
pyzmq==22.1.0
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rich==10.12.0
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
scikit-learn==1.0
scipy==1.6.3
seaborn==0.11.1
shortuuid==1.0.1
shtab==1.4.2
six==1.15.0
sklearn==0.0
smart-open==5.1.0
smmap==4.0.0
soupsieve==2.2.1
tables==3.6.1
tabulate==0.8.9
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==2.2.0
toml==0.10.2
tornado==6.1
tqdm==4.61.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.5
voluptuous==0.12.2
wcwidth==0.2.5
websocket-client==1.1.0
Werkzeug==2.0.1
wrapt==1.12.1
xmltodict==0.12.0
yacs==0.1.8
yarl==1.7.0
youtube-dl==2021.6.6
zc.lockfile==2.0
zipp==3.4.1
